Refactor main app into mixins
