+++ src/Main_App/file_selection.py
@@ -1,0 +1,63 @@
+# -*- coding: utf-8 -*-
+"""File and folder selection mixin."""
+import os
+import glob
+from tkinter import filedialog, messagebox
+
+class FileSelectionMixin:
+    def select_save_folder(self):
+        folder = filedialog.askdirectory(title="Select Parent Folder for Excel Output")
+        if folder:
+            self.save_folder_path.set(folder)
+            self.log(f"Output folder: {folder}")
+        else:
+            self.log("Save folder selection cancelled.")
+
+    def update_select_button_text(self):
+        """Update the label on the Select EEG File… button to match the current mode."""
+        if self.file_mode.get() == "Single":
+            self.select_button.configure(text="Select EEG File…")
+        else:
+            self.select_button.configure(text="Select Folder…")
+
+    def select_data_source(self):
+        self.data_paths = []
+        file_ext = "*" + self.file_type.get().lower()
+        file_type_desc = self.file_type.get().upper()
+        try:
+            if self.file_mode.get() == "Single":
+                ftypes = [(f"{file_type_desc} files", file_ext)]
+                other_ext = "*.set" if file_type_desc == ".BDF" else "*.bdf"
+                other_desc = ".SET" if file_type_desc == ".BDF" else ".BDF"
+                ftypes.append((f"{other_desc} files", other_ext))
+                ftypes.append(("All files", "*.*"))
+
+                file_path = filedialog.askopenfilename(title="Select EEG File", filetypes=ftypes)
+                if file_path:
+                    selected_ext = os.path.splitext(file_path)[1].lower()
+                    if selected_ext in ['.bdf', '.set']:
+                        self.file_type.set(selected_ext.upper())
+                        self.log(f"File type set to {selected_ext.upper()}")
+                    self.data_paths = [file_path]
+                    self.log(f"Selected file: {os.path.basename(file_path)}")
+                else:
+                    self.log("No file selected.")
+            else:
+                folder = filedialog.askdirectory(title=f"Select Folder with {file_type_desc} Files")
+                if folder:
+                    search_path = os.path.join(folder, file_ext)
+                    found_files = sorted(glob.glob(search_path))
+                    if found_files:
+                        self.data_paths = found_files
+                        self.log(f"Selected folder: {folder}, Found {len(found_files)} '{file_ext}' file(s).")
+                    else:
+                        self.log(f"No '{file_ext}' files found in {folder}.")
+                        messagebox.showwarning("No Files Found", f"No '{file_type_desc}' files found in:\n{folder}")
+                else:
+                    self.log("No folder selected.")
+        except Exception as e:
+            self.log(f"Error selecting data: {e}")
+            messagebox.showerror("Selection Error", f"Error during selection:\n{e}")
+
+        self._max_progress = len(self.data_paths) if self.data_paths else 1
+        self.progress_bar.set(0)
